refactor: Extract Checkbox.jsx
