+++ frontend/src/components/README.md
@@ -1,0 +1,5 @@
+# Components
+
+A set of general UI components.
+
+Many are "connected" tools which use a "link" prop to connect to the app state.
