+++ frontend/src/utils.jsx
@@ -8,27 +8,6 @@ import App from "./app";
  * through a "link" prop to connect to the app state
  */
 
-export const Checkbox = ({link, text, side, onChange, ...rest}) => (
-  <div>
-    {side === "right" ? text : ""}
-    <input
-      {...rest}
-      type="checkbox"
-      onChange={onChange || function (e) {
-        App.save(link, e.currentTarget.checked);
-      }}
-      checked={App.state[link]}/>
-    {side === "left" ? text : ""}
-  </div>
-);
-
-Checkbox.propTypes = {
-  link: PropTypes.string,
-  text: PropTypes.string,
-  side: PropTypes.string,
-  onChange: PropTypes.func
-};
-
 export const Spaced = ({elements}) => (
   elements
     .map((x, index) => <span key={index}>{x}</span>)
